# openapi_tramites.json
{"openapi":"3.1.0","info":{"title":"Gestión de Trámites de Servicio","description":"API RESTful para gestionar solicitudes de alta y baja de un servicio.","version":"2.0.0"},"paths":{"/v1/admin/login":{"post":{"tags":["Tokens"],"summary":"Login Administrador","operationId":"login_administrador_v1_admin_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/tramites":{"get":{"tags":["Trámites"],"summary":"Listar Tramites","operationId":"listar_tramites_v1_tramites_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"estado","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/EstadoTramite"},{"type":"null"}],"title":"Estado"}},{"name":"tipo","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/TipoTramite"},{"type":"null"}],"title":"Tipo"}},{"name":"orden","in":"query","required":false,"schema":{"type":"string","default":"desc","title":"Orden"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","default":1,"title":"Page"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":20,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TramiteAdminResponse"},"title":"Response Listar Tramites V1 Tramites Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Trámites"],"summary":"Crear Tramite","operationId":"crear_tramite_v1_tramites_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TramiteRequest"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TramiteCreadoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/tramites/{id}":{"get":{"tags":["Trámites"],"summary":"Obtener Tramite","operationId":"obtener_tramite_v1_tramites__id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"anyOf":[{"$ref":"#/components/schemas/TramitePublicoResponse"},{"$ref":"#/components/schemas/TramiteAdminResponse"}],"title":"Response Obtener Tramite V1 Tramites  Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["Trámites"],"summary":"Actualizar Estado Tramite","operationId":"actualizar_estado_tramite_v1_tramites__id__patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActualizarEstadoRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TramiteAdminResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/archivos":{"post":{"tags":["Archivos"],"summary":"Subir Archivo","operationId":"subir_archivo_v1_archivos_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_subir_archivo_v1_archivos_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArchivoResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ActualizarEstadoRequest":{"properties":{"estado":{"$ref":"#/components/schemas/EstadoTramite"},"motivo-rechazo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Motivo-Rechazo"},"nota-interna":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nota-Interna"}},"type":"object","required":["estado"],"title":"ActualizarEstadoRequest"},"ArchivoEnTramite":{"properties":{"id":{"type":"string","title":"Id"},"nombre":{"type":"string","title":"Nombre"},"url":{"type":"string","title":"Url"}},"type":"object","required":["id","nombre","url"],"title":"ArchivoEnTramite"},"ArchivoResponse":{"properties":{"id":{"type":"string","title":"Id"},"nombre-archivo":{"type":"string","title":"Nombre-Archivo"},"tipo-mime":{"type":"string","title":"Tipo-Mime"},"tamano-bytes":{"type":"integer","title":"Tamano-Bytes"},"url":{"type":"string","title":"Url"},"fecha-subida":{"type":"string","format":"date-time","title":"Fecha-Subida"}},"type":"object","required":["id","nombre-archivo","tipo-mime","tamano-bytes","url","fecha-subida"],"title":"ArchivoResponse"},"Body_subir_archivo_v1_archivos_post":{"properties":{"archivo":{"type":"string","contentMediaType":"application/octet-stream","title":"Archivo"}},"type":"object","required":["archivo"],"title":"Body_subir_archivo_v1_archivos_post"},"EstadoTramite":{"type":"string","enum":["PENDIENTE","APROBADO","RECHAZADO"],"title":"EstadoTramite"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"usuario":{"type":"string","title":"Usuario"},"clave":{"type":"string","title":"Clave"}},"type":"object","required":["usuario","clave"],"title":"LoginRequest"},"LoginResponse":{"properties":{"token":{"type":"string","title":"Token"},"tipo":{"type":"string","title":"Tipo","default":"Bearer"},"expira-en":{"type":"integer","title":"Expira-En"}},"type":"object","required":["token","expira-en"],"title":"LoginResponse"},"TipoTramite":{"type":"string","enum":["ALTA","BAJA"],"title":"TipoTramite"},"TramiteAdminResponse":{"properties":{"id":{"type":"string","title":"Id"},"tipo":{"$ref":"#/components/schemas/TipoTramite"},"estado":{"$ref":"#/components/schemas/EstadoTramite"},"nombre":{"type":"string","title":"Nombre"},"apellido":{"type":"string","title":"Apellido"},"dni":{"type":"string","title":"Dni"},"mail":{"type":"string","format":"email","title":"Mail"},"telefono":{"type":"string","title":"Telefono"},"direccion":{"type":"string","title":"Direccion"},"fecha-creacion":{"type":"string","format":"date-time","title":"Fecha-Creacion"},"fecha-resolucion":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Fecha-Resolucion"},"motivo-rechazo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Motivo-Rechazo"},"nota-interna":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nota-Interna"},"resuelto-por":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Resuelto-Por"},"archivos":{"items":{"$ref":"#/components/schemas/ArchivoEnTramite"},"type":"array","title":"Archivos","default":[]}},"type":"object","required":["id","tipo","estado","nombre","apellido","dni","mail","telefono","direccion","fecha-creacion"],"title":"TramiteAdminResponse"},"TramiteCreadoResponse":{"properties":{"id":{"type":"string","title":"Id"},"tipo":{"$ref":"#/components/schemas/TipoTramite"},"estado":{"$ref":"#/components/schemas/EstadoTramite"},"mensaje":{"type":"string","title":"Mensaje","default":"Su trámite fue registrado. Recibirá un mail con su ID de seguimiento."}},"type":"object","required":["id","tipo","estado"],"title":"TramiteCreadoResponse"},"TramitePublicoResponse":{"properties":{"id":{"type":"string","title":"Id"},"tipo":{"$ref":"#/components/schemas/TipoTramite"},"estado":{"$ref":"#/components/schemas/EstadoTramite"},"fecha-creacion":{"type":"string","format":"date-time","title":"Fecha-Creacion"},"detalle-estado":{"type":"string","title":"Detalle-Estado"},"motivo-rechazo":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Motivo-Rechazo"}},"type":"object","required":["id","tipo","estado","fecha-creacion","detalle-estado"],"title":"TramitePublicoResponse"},"TramiteRequest":{"properties":{"tipo":{"$ref":"#/components/schemas/TipoTramite"},"nombre":{"type":"string","title":"Nombre"},"apellido":{"type":"string","title":"Apellido"},"dni":{"type":"string","title":"Dni"},"mail":{"type":"string","format":"email","title":"Mail"},"telefono":{"type":"string","title":"Telefono"},"direccion":{"type":"string","title":"Direccion"},"archivo_id_1":{"type":"string","title":"Archivo Id 1"},"archivo_id_2":{"type":"string","title":"Archivo Id 2"}},"type":"object","required":["tipo","nombre","apellido","dni","mail","telefono","direccion","archivo_id_1","archivo_id_2"],"title":"TramiteRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}